Landing Page
Landing page, including navbar, and css for it. A mock up of a gallery
page as I came across code for it whilst looking for other stuff. Note:
There's also something you can do with links and opening them in a frame
inside the original page rather than in a new tab which might be
something to look at in the future as it may enable blog-post-like
displays or even a frame for comparision on another biography page.